Update main.cpp
Single Cycle to Pipeline
I changed a lot for this and dont know exactly what lines so I did the whole cut and paste thing again.

changed jumps and branches to alter PC in execute stage. I believe everything works and with some verification we can move on to Forwarding.